refactor: split SettingsPage into tab components + add time capsule Rust tests
- Extract 8 tab components to src/components/settings/tabs/:
  GeneralTab (STT, reminders, time capsule, appearance)
  PrivacyTab (2FA modals, reset confirm, biometric unlock, data stats)
  SyncTab (WebDAV config)
  AITab (OpenAI/Ollama config)
  HealthTab (Oura Ring)
  ExportTab (selective export)
  AboutTab (log level, version, updates)
  index.ts + types.ts barrel/type exports
- SettingsPage.tsx: 2239 → 744 lines (coordinator shell only)
  Retains: tab nav, scroll-to-section, tab-switch data loading,
  export password modal + rate limiting, refresh2FAStatus
- Add 6 #[cfg(test)] Rust tests to time_capsule.rs:
  seal_entry sets columns, rejects past dates, double-seal guard
  unseal_entry clears columns + defaults capsule_type to 'anniversary'
  get_due_capsules returns past-due, excludes today M-D anniversary
