+++ src/components/settings/tabs/AboutTab.tsx
@@ -1,0 +1,154 @@
+import type { AppSettings } from '../../../types/settings';
+import { SettingSection } from '../index';
+import { UpdatePanel } from '../../updater/UpdatePanel';
+import type { UseUpdateCheckReturn } from '../../../hooks/useUpdateCheck';
+import { invoke } from '@tauri-apps/api/core';
+import { logger } from '../../../lib/services/logger';
+import type { LogLevel } from '../../../lib/services/logger';
+
+interface AboutTabProps {
+  settings: AppSettings;
+  updateHook: UseUpdateCheckReturn;
+  appVersion: string;
+  logPath: string | null;
+  handleLogLevelChange: (level: LogLevel) => void;
+}
+
+export function AboutTab({
+  settings,
+  updateHook,
+  appVersion,
+  logPath,
+  handleLogLevelChange,
+}: AboutTabProps) {
+  return (
+    <div id="panel-about" role="tabpanel" className="space-y-6">
+
+      {/* Updates section */}
+      <SettingSection
+        title="Updates"
+        description="Keep MoodHaven Journal up to date"
+      >
+        <UpdatePanel hook={updateHook} currentVersion={appVersion} />
+      </SettingSection>
+
+      <SettingSection
+        title="About MoodHaven Journal"
+        description="App information and credits"
+      >
+        <div className="space-y-4">
+          <div className="flex items-center justify-between py-3 border-b border-slate-100 dark:border-slate-700">
+            <p className="text-slate-700 dark:text-slate-200">App Version</p>
+            <p className="text-slate-500 dark:text-slate-400 font-mono bg-slate-100 dark:bg-slate-800 px-2 py-0.5 rounded">
+              v{appVersion}
+            </p>
+          </div>
+
+          <div className="flex items-center justify-between py-3 border-b border-slate-100 dark:border-slate-700">
+            <p className="text-slate-700 dark:text-slate-200">Settings Version</p>
+            <p className="text-slate-500 dark:text-slate-400 font-mono bg-slate-100 dark:bg-slate-800 px-2 py-0.5 rounded">
+              {settings.version}
+            </p>
+          </div>
+
+          <div className="flex items-center justify-between py-3 border-b border-slate-100 dark:border-slate-700">
+            <p className="text-slate-700 dark:text-slate-200">Platform</p>
+            <p className="text-slate-500 dark:text-slate-400">
+              {navigator.platform}
+            </p>
+          </div>
+
+          <div className="flex items-center justify-between py-3 border-b border-slate-100 dark:border-slate-700">
+            <div>
+              <p className="text-slate-700 dark:text-slate-200">Log Level</p>
+              <p className="text-xs text-slate-400 dark:text-slate-500 mt-0.5">Debug is verbose — use only for troubleshooting</p>
+            </div>
+            <select
+              aria-label="Log level"
+              value={settings.logLevel ?? 'warn'}
+              onChange={(e) => handleLogLevelChange(e.target.value as LogLevel)}
+              className="px-3 py-1 text-sm rounded-lg bg-slate-100 dark:bg-slate-700 text-slate-700 dark:text-slate-300 border-0 cursor-pointer"
+            >
+              <option value="error">Error</option>
+              <option value="warn">Warn</option>
+              <option value="info">Info</option>
+              <option value="debug">Debug</option>
+            </select>
+          </div>
+
+          <div className="flex items-center justify-between py-3 border-b border-slate-100 dark:border-slate-700">
+            <p className="text-slate-700 dark:text-slate-200">Log File</p>
+            <button
+              onClick={() => {
+                if (logPath) {
+                  invoke('open_log_folder').catch((e: unknown) => {
+                    logger.error('open_log_folder failed', { err: String(e) });
+                  });
+                }
+              }}
+              disabled={!logPath}
+              className="px-3 py-1 text-sm rounded-lg bg-slate-100 dark:bg-slate-700 text-slate-700 dark:text-slate-300 hover:bg-slate-200 dark:hover:bg-slate-600 disabled:opacity-40 disabled:cursor-not-allowed transition-colors"
+            >
+              Open Log Folder
+            </button>
+          </div>
+
+          <div className="pt-4">
+            <p className="text-sm text-slate-600 dark:text-slate-400 leading-relaxed">
+              MoodHaven Journal is a privacy-focused mood tracking and journaling application.
+              All your data is stored locally on your device and encrypted using
+              industry-standard AES-256-GCM encryption.
+            </p>
+            <p className="text-sm text-slate-500 dark:text-slate-400 leading-relaxed mt-2">
+              MoodHaven Journal is built by one person. If it brings value to your life, a coffee goes a long way.
+            </p>
+          </div>
+
+          <div className="p-4 bg-gradient-to-br from-violet-50 to-purple-50 dark:from-violet-900/20 dark:to-purple-900/20 rounded-xl border border-violet-100 dark:border-violet-800">
+            <p className="text-sm font-medium text-violet-700 dark:text-violet-300 mb-2">
+              Built with
+            </p>
+            <div className="flex flex-wrap gap-2">
+              {['Tauri', 'React', 'TypeScript', 'TailwindCSS', 'Rust', 'SQLite'].map((tech) => (
+                <span
+                  key={tech}
+                  className="px-2 py-1 text-xs font-medium bg-white dark:bg-slate-800 text-violet-600 dark:text-violet-400 rounded-md shadow-sm"
+                >
+                  {tech}
+                </span>
+              ))}
+            </div>
+          </div>
+
+          {/* Links */}
+          <div className="flex flex-wrap gap-3 pt-2">
+            <a
+              href="https://github.com/kenlacroix/moodhaven-journal#readme"
+              target="_blank"
+              rel="noopener noreferrer"
+              className="flex items-center gap-1.5 px-4 py-2 text-sm text-violet-600 dark:text-violet-400 border border-violet-200 dark:border-violet-800 rounded-xl hover:bg-violet-50 dark:hover:bg-violet-900/20 transition-colors"
+            >
+              User Guide ↗
+            </a>
+            <a
+              href="https://github.com/kenlacroix/moodhaven-journal/issues"
+              target="_blank"
+              rel="noopener noreferrer"
+              className="flex items-center gap-1.5 px-4 py-2 text-sm text-slate-600 dark:text-slate-400 border border-slate-200 dark:border-slate-700 rounded-xl hover:bg-slate-50 dark:hover:bg-slate-800 transition-colors"
+            >
+              Share Feedback ↗
+            </a>
+            <a
+              href="https://buymeacoffee.com/moodbloom"
+              target="_blank"
+              rel="noopener noreferrer"
+              className="flex items-center gap-1.5 px-4 py-2 text-sm text-slate-600 dark:text-slate-400 border border-slate-200 dark:border-slate-700 rounded-xl hover:bg-slate-50 dark:hover:bg-slate-800 transition-colors"
+            >
+              Buy Me a Coffee ↗
+            </a>
+          </div>
+        </div>
+      </SettingSection>
+    </div>
+  );
+}
