+++ src/components/settings/tabs/index.ts
@@ -1,0 +1,7 @@
+export { GeneralTab } from './GeneralTab';
+export { PrivacyTab } from './PrivacyTab';
+export { SyncTab } from './SyncTab';
+export { AITab } from './AITab';
+export { HealthTab } from './HealthTab';
+export { ExportTab } from './ExportTab';
+export { AboutTab } from './AboutTab';
